backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
scikit-learn
pandas
numpy
joblib
